requirements.txt
alembic==1.8.1
appdirs==1.4.4
argcomplete==2.0.0
click==8.1.3
Flask==2.2.2
Flask-Migrate==4.0.0
Flask-SQLAlchemy==3.0.2
greenlet==2.0.1
importlib-metadata==5.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.2.3
MarkupSafe==2.1.1
Pillow==9.3.0
psycopg2-binary==2.9.5
pyserial==3.5
python-dotenv==0.21.0
python-escpos==2.2.0
pyusb==1.2.1
PyYAML==6.0
qrcode==7.3.1
six==1.16.0
SQLAlchemy==1.4.44
tabulate==0.9.0
Werkzeug==2.2.2
zipp==3.10.0
